# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/houseboats/{houseboatId}/my-settings": {
      "get": {
        "operationId": "RbacController_mySettings",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The caller's own preferences on a boat (notification toggles).\n\nDeliberately NOT decorated with @RequirePermission: a member with no\n`settings` permission must still be able to manage their own notifications.\nBecause PermissionGuard skips undecorated routes, this handler does its own\nmembership check — without it the route would be authenticated but\nunscoped, and anyone could read any boat's membership row.",
        "tags": [
          "Rbac"
        ]
      },
      "patch": {
        "operationId": "RbacController_updateMySettings",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MySettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/roles": {
      "get": {
        "operationId": "RbacController_listRoles",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      },
      "post": {
        "operationId": "RbacController_createRole",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/roles/{roleId}": {
      "patch": {
        "operationId": "RbacController_updateRole",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/members": {
      "get": {
        "operationId": "RbacController_listMembers",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Rbac"
        ]
      },
      "post": {
        "operationId": "RbacController_addMember",
        "parameters": [
          {
            "name": "houseboatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/members/{membershipId}/role": {
      "patch": {
        "operationId": "RbacController_changeRole",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
    "/api/houseboats/{houseboatId}/members/{membershipId}/exit": {
      "post": {
        "operationId": "RbacController_exitMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rbac"
        ]
      }
    },
...
  },
  "info": {
    "title": "Houseboat API",
    "description": "Booking SaaS backend",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hb_access"
      }
    },
    "schemas": {
...
      "MySettingsDto": {
        "type": "object",
        "properties": {
          "notificationPrefs": {
            "type": "object",
            "description": "Per-event notification toggles: {\"booking\":true,\"payment_due\":false,…}.\nLeft open rather than a fixed DTO so adding an event type needs no\nmigration; the console owns which keys it renders."
          }
        }
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "permissions": {
            "type": "object"
          },
          "isTemplate": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "permissions": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 64
          },
          "roleId": {
            "type": "string",
            "maxLength": 64
          },
          "shareholderPct": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "phone",
          "roleId"
        ]
      },
      "ChangeRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "roleId"
        ]
      },
...
    }
  }
}